requirements-pipreq.txt
benepar==0.2.0
flair==0.11.3
matplotlib==3.5.3
networkx==2.6.3
nltk==3.7
numpy==1.21.6
pandas==1.3.5
PyYAML==6.0
Requests==2.31.0
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
serpyco==1.3.9
signac==1.7.0
six==1.16.0
spacy==3.2.5
torch==1.10.2+cu111
tqdm==4.63.2
pre-commit==2.17.0
transformers==4.29.2
